conseguido, evaluaciones con alumno y nota por asignatura

+++ Etapa1/Entidades/Evaluacion.cs
@@ -11,7 +11,7 @@ namespace CoreEscuela.Entidades
         
         public Asignatura Asignatura { get; set; }
 
-        public float Nota { get; set; }
+        public double Nota { get; set; }
 
         public Evaluacion()
         {
